Feat/mood service (#4)
* (feat): Create mood service and controller, WIP

* (wip): Friends feature x moods

* (feat): Unit testing + check in friedns table for moods /friends/<id>

* (feat): Cargo fmt

* (fix): Cargo clippy

+++ migrations/005_create_moods_table/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE moods (
     id SERIAL PRIMARY KEY NOT NULL,
-    user_id BIGINT NOT NULL REFERENCES users(id),
-    game_id BIGINT NOT NULL REFERENCES games(id),
+    user_id INTEGER NOT NULL REFERENCES users(id),
+    game_id INTEGER NOT NULL REFERENCES games(id),
     begin_timestamp TIMESTAMP WITH TIME ZONE NOT NULL,
     end_timestamp TIMESTAMP WITH TIME ZONE NOT NULL
 );
